- Small refactoring - Fixed shovel bug out of bounds - Moved backyard to model package
